pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "starship"
version = "1.0.0"
description = "AI Stars GitHub Knowledge Engine - Transform your GitHub stars into an intelligent knowledge base"
authors = [
    {name = "Starship Team", email = "zmzposa@outlook.com"},
]
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "github", "knowledge-base", "rag", "vector-search", "autogen"]
dependencies = [
    # Core dependencies
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    # Database
    "aiosqlite>=0.19.0",
    # HTTP client
    "httpx>=0.25.0",
    # Scheduling
    "apscheduler>=3.10.0",
    # Existing dependencies
    "autogen-agentchat>=0.6.1",
    "autogen-ext[ollama,chromadb]>=0.4.6",
    "mcp[cli]>=1.0.0",
    "chromadb>=0.4.0",
    "loguru>=0.7.3",
    "sentence-transformers>=5.0.0",
    "orjson>=3.11.1",
    "polars>=1.32.0",
    "openai>=2.14.0",
    "progress>=1.6.1",
]

[project.optional-dependencies]
dev = [
    "pyyaml>=6.0.2",  # Only used in tests
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/dustyposa/starship"
Documentation = "https://starship-ai.github.io/starship"
Repository = "https://github.com/dustyposa/starship"
"Bug Tracker" = "https://github.com/dustyposa/starship/issues"

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple/"

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"

[project.scripts]
starship = "starship.cli:main"
starship-server = "src.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.15.1",
]
